feat: retire loaf change and loaf spec; freeze the corpus as history
The two legacy write surfaces come out of the CLI: dispatch, help,
agent help, and reference entries for loaf change and loaf spec, the
Change scaffolder and its templates, the Change list/tasks/state/
evidence/provenance/conversion commands, and the spec command
namespace. Invoking either now fails as an unknown command.

The read model survives the writers: journal context still derives its
active-changes and active-lineage layers from the frozen corpus, the
flag-based release path keeps its prepared-tree helpers, SQLite spec
records stay reachable through loaf trace and loaf state export spec,
and loaf task and loaf intent remain as legacy read surfaces.
Remediation text now points at those read paths instead of the retired
writers.

docs/changes/ is byte-for-byte unchanged and never written again — the
27 Change folders are history, deliberately unconverted. No hooks
enforced Change structure, so loaf check's hook roster is unchanged.
The CLI reference regenerates and distribution artifacts rebuild for
the reduced surface.
