+++ content/skills/loaf-reference/SKILL.md
@@ -19,7 +19,7 @@ The Loaf operating manual for agents: how to discover commands, diagnose project
 ## Operating Rules
 
 - Get exact, current syntax live: `loaf --help` lists every command, `loaf <command> --help` details one. This index is a map, not the contract.
-- Prefer `--json` surfaces when diagnosing: `loaf config check --json`, `loaf state doctor --json`, `loaf change check --json`. Parse the structured output instead of scraping human-readable text.
+- Prefer `--json` surfaces when diagnosing: `loaf config check --json`, `loaf state doctor --json`. Parse the structured output instead of scraping human-readable text.
 - Run the deterministic CLI command before hand-editing anything it manages; the command owns its files.
 - Use `--fix` only for safe, mechanical repairs, and review what it changed.
 - Ask the user for project-owned choices — GitHub account, tracker or integration election, which harnesses to install — never guess them.
@@ -61,14 +61,12 @@ Names and one-line purposes only. Run `loaf <command> --help` for options, argum
 | `loaf release` | Create a new release with changelog, version bump, and tag | suggest, cut |
 | `loaf search` | Search SQLite artifact bodies, journal entries, and indexed docs | — |
 | `loaf docs` | Manage docs/ indexing | index |
-| `loaf change` | Shape-first Change artifacts: git-canonical work context under docs/changes/ | init, check, list |
 | `loaf render` | Maintain committed durable Markdown renders | sweep |
 | `loaf state` | Manage native SQLite state | path, status, init, doctor, repair legacy-project-database, repair relationship-origin, repair journal-search, migrate markdown, migrate storage-home, migrate schema, migrate deferrals, migrate lifecycle-statuses, migrate alias-orphans, migrate journal-duplicates, migrate journal-first, backup, backup verify, backup restore, restore-ephemerals, verify-ephemerals, export, export all, export triage, export spec, export release-readiness |
 | `loaf journal` | Record and read the project-scoped journal (the durable record across all conversations) | log, recent, search, show, context, export, defer |
 | `loaf project` | Manage durable project identity | list, show, identity, rename, move, delete |
 | `loaf migrate` | Run native migration workflows | markdown, storage-home, schema, lifecycle-statuses, journal-first, worktree-storage |
 | `loaf task` | Manage project tasks; superseded by loaf issue for new work | list, show, status, create, update, archive, refresh, sync |
-| `loaf spec` | Manage project specs; superseded by loaf issue for new work | new, edit, list, show, status, render, finalize, archive, delete |
 | `loaf issue` | Manage issues in native SQLite state | new, show, list, tree, frontier, start, stop, edit, status, dod, dod add, dod list, dod remove, dod claim, dod unclaim, promote, check, verify, bucket, link, render, export, pull, push, reconcile |
 | `loaf report` | Manage durable reports (research, audits, investigations) | list, show, render, generate, create, edit, finalize, archive |
 | `loaf finding` | Manage report findings and verdicts in native SQLite state | list, show, create, verdict, import-json |
