+++ internal/cli/authority_test.go
@@ -19,7 +19,6 @@ func TestCommandAuthorityForUsesExplicitBasicLeafPrefixes(t *testing.T) {
 		{"report", "generate", "triage", "--format", "markdown"},
 		{"report", "generate", "release-readiness", "--format", "markdown"},
 		{"project", "identity"},
-		{"spec", "status", "SPEC-001", "done"},
 		{"issue", "list"},
 		{"issue", "show", "LOAF-1"},
 		{"issue", "tree"},
@@ -59,18 +58,13 @@ func TestCommandAuthorityForDefaultsToOperatorForParentsAndUnsafeLeaves(t *testi
 		{"brainstorm", "capture", "--title", "brainstorm"},
 		{"idea", "capture", "--title", "idea"},
 		{"project", "rename", "new-name"},
-		{"spec", "finalize", "SPEC-001"},
-		{"spec", "new", "slug", "--body-file", "body.md"},
 		{"issue", "new", "title", "--body", "body"},
 		{"issue", "edit", "LOAF-1", "--message", "body"},
 		{"issue", "status", "LOAF-1", "cancelled"},
 		{"issue", "start", "LOAF-1"},
 		{"issue", "stop", "LOAF-1"},
-		{"spec", "delete", "SPEC-001", "--yes"},
 		{"kb", "review", "docs/knowledge/example.md"},
 		{"kb", "glossary", "upsert", "term"},
-		{"change", "init", "new-change"},
-		{"change", "check"},
 		{"release"},
 		{"not-a-command"},
 	} {
