--- internal/cli/change_brief_template.md
@@ -1,42 +1,0 @@
-<!-- brief.md is the optional archeological kickstart — the original unshaped ask.
-     May accrete parked problem-space concepts while the change is captured; freezes when shape.md exists.
-     Superseded by shape.md; never mechanically load-bearing.
-     A brief-only folder is legal and non-executable (captured, not shaped). -->
-
-# [Brief title]
-
-## Problem Statement
-
-[What friction, gap, or unmet need exists? Be specific about the pain — vague problems produce vague solutions. Problem-space only; do not design the approach here.]
-
-## Who Has It
-
-[Who experiences this problem? Role, context, and how often the pain shows up. Avoid unqualified "users" or "developers."]
-
-## Current Alternatives
-
-[What do they do today? Existing tools, manual workarounds, or "nothing" are all valid. Understanding the status quo clarifies what better means.]
-
-## Value Proposition
-
-[Why is solving this worth it? What becomes true for the people who have the problem if this lands? Describe value, not features or architecture.]
-
-## Constraints
-
-[Non-negotiable bounds: technical, legal, organizational, or philosophical. Things that limit the solution space before design begins.]
-
-- [Constraint 1]
-
-## Sequencing and Relationships
-
-[How this relates to other work — series order, release cohort, dependencies stated as prose. No machine relation fields; narrative order only.]
-
-## Sources and Research Links
-
-[Evidence that informed this framing — competitive scans, research notes, issue links, prior art. Link rather than paste.]
-
-## Open Questions
-
-[Unresolved problem-space items. Mark urgency: blocking (must resolve before shaping) or deferrable.]
-
-- [ ] [Question] — blocking | deferrable
