--- internal/cli/change_design_template.md
@@ -1,20 +1,0 @@
-<!-- design.md is the optional design surface for UI, protocol, or schema detail
-     that would crowd shape.md or plan.md. Accretive; not load-bearing for the gate. -->
-
-# Design — [Change Title]
-
-## Intent
-
-[What this design clarifies that shape.md does not.]
-
-## Surfaces
-
-[Screens, APIs, schemas, or protocols under design.]
-
-## Decisions
-
-1. **[Decision.]** [Rationale.]
-
-## Open questions
-
-- [What remains undecided.]
