+++ internal/cli/change_json_test.go
@@ -108,240 +108,3 @@ func TestAssembleLegacyLoadsTargetRelease(t *testing.T) {
 		t.Fatalf("node = %+v", node)
 	}
 }
-
-func TestDeriveChangeCohorts(t *testing.T) {
-	nodes := []changeNode{
-		{Slug: "a", TargetRelease: "2.0.0", Folder: "docs/changes/20260710-a"},
-		{Slug: "b", TargetRelease: "2.0.0", Folder: "docs/changes/20260711-b"},
-		{Slug: "c", TargetRelease: "2.1.0", Folder: "docs/changes/20260712-c"},
-		{Slug: "d", Folder: "docs/changes/20260713-d"},
-	}
-	cohorts := deriveChangeCohorts(nodes)
-	if len(cohorts["2.0.0"]) != 2 || len(cohorts["2.1.0"]) != 1 || len(cohorts[""]) != 0 {
-		t.Fatalf("cohorts = %#v", cohorts)
-	}
-}
-
-func TestChangeCheckFailsClosedOnMalformedJSONBesideMarkdown(t *testing.T) {
-	repo := initCLIGitRepo(t)
-	folder := writeChangeFolder(t, repo, "20260727-example", changeDoc(changeFrontmatter("example", "2026-07-27", "example"), append(productSections(), executableSections()...)...))
-	if err := os.WriteFile(filepath.Join(folder, "change.json"), []byte(`{bad`), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	out, err := runChangeCheckJSON(t, repo, "docs/changes/20260727-example")
-	if err == nil || !findingsContain(out.Findings, "malformed change.json") {
-		t.Fatalf("err=%v out=%+v", err, out)
-	}
-}
-
-func TestDeriveChangeRetargetEventsUnionsSurfaces(t *testing.T) {
-	repo := initCLIGitRepo(t)
-	folder := writeChangeFolder(t, repo, "20260727-example", changeDoc(
-		"---\nchange: example\ncreated: 2026-07-27\nbranch: example\ntarget_release: 2.0.0\n---\n",
-		append(productSections(), executableSections()...)...,
-	))
-	commitAllChangeTest(t, repo, "docs: add targeted change")
-
-	jsonPath := filepath.Join(folder, "change.json")
-	if err := os.WriteFile(jsonPath, []byte(`{"change":"example","created":"2026-07-27","branch":"example","target_release":"2.1.0"}`), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.Remove(filepath.Join(folder, "change.md")); err != nil {
-		t.Fatal(err)
-	}
-	commitAllChangeTest(t, repo, "docs: convert and retarget")
-
-	events, err := deriveChangeRetargetEvents(repo, commandOutput)
-	if err != nil {
-		t.Fatal(err)
-	}
-	found := false
-	for _, event := range events {
-		if event.Folder == "docs/changes/20260727-example" && event.From == "2.0.0" && event.To == "2.1.0" {
-			found = true
-		}
-	}
-	if !found {
-		t.Fatalf("events = %#v, want 2.0.0 -> 2.1.0", events)
-	}
-}
-
-func TestRetentionAllowsAtomicConversionAndBlocksJSONTargetDeletion(t *testing.T) {
-	repo := initCLIGitRepo(t)
-	folder := writeChangeFolder(t, repo, "20260727-example", changeDoc(
-		"---\nchange: example\ncreated: 2026-07-27\nbranch: example\ntarget_release: 2.0.0\n---\n",
-		append(productSections(), executableSections()...)...,
-	))
-	commitAllChangeTest(t, repo, "docs: add targeted change")
-
-	if err := os.WriteFile(filepath.Join(folder, "change.json"), []byte(`{"change":"example","created":"2026-07-27","branch":"example","target_release":"2.0.0"}`), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.Remove(filepath.Join(folder, "change.md")); err != nil {
-		t.Fatal(err)
-	}
-	commitAllChangeTest(t, repo, "docs: atomic convert")
-
-	deleted, err := deletedLineageChangesWithOutput(repo, commandOutput)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(deleted) != 0 {
-		t.Fatalf("atomic conversion treated as deletion: %v", deleted)
-	}
-
-	if err := os.Remove(filepath.Join(folder, "change.json")); err != nil {
-		t.Fatal(err)
-	}
-	commitAllChangeTest(t, repo, "docs: delete target-declaring json")
-	deleted, err = deletedLineageChangesWithOutput(repo, commandOutput)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(deleted) == 0 {
-		t.Fatal("expected retention finding for deleted target-declaring change.json")
-	}
-}
-
-func TestResolveChangeFolderFindsJSONLayout(t *testing.T) {
-	repo := initCLIGitRepo(t)
-	folder := filepath.Join(repo, "docs", "changes", "20260727-example")
-	if err := os.MkdirAll(folder, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(folder, "change.json"), []byte(`{"change":"example","created":"2026-07-27","branch":"example"}`), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	gotFolder, gotFile, err := resolveChangeFolder(repo, "docs/changes/20260727-example")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if filepath.Base(gotFolder) != "20260727-example" || filepath.Base(gotFile) != "change.json" {
-		t.Fatalf("folder=%q file=%q", gotFolder, gotFile)
-	}
-}
-
-func taskPacketBody(change, id, title string, checked bool) string {
-	mark := " "
-	if checked {
-		mark = "x"
-	}
-	return "---\nchange: " + change + "\nid: " + id + "\ntitle: " + title + "\n---\n\n# " + id + " — " + title + "\n\n## Steps\n\n- [" + mark + "] Do the work\n"
-}
-
-func atomicConvertFolder(t *testing.T, repo, folder, slug string, checked bool) {
-	t.Helper()
-	dir := filepath.Join(repo, "docs", "changes", folder)
-	if err := os.MkdirAll(filepath.Join(dir, "tasks"), 0o755); err != nil {
-		t.Fatalf("MkdirAll tasks: %v", err)
-	}
-	meta := "{\n  \"change\": \"" + slug + "\",\n  \"created\": \"2026-07-27\",\n  \"branch\": \"" + slug + "\",\n  \"target_release\": \"2.0.0\"\n}\n"
-	if err := os.WriteFile(filepath.Join(dir, "change.json"), []byte(meta), 0o644); err != nil {
-		t.Fatalf("WriteFile change.json: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(dir, "shape.md"), []byte(authoredShapeBody()), 0o644); err != nil {
-		t.Fatalf("WriteFile shape.md: %v", err)
-	}
-	taskName := "TASK-001-do-work.md"
-	if err := os.WriteFile(filepath.Join(dir, "tasks", taskName), []byte(taskPacketBody(slug, "TASK-001", "Do work", checked)), 0o644); err != nil {
-		t.Fatalf("WriteFile task: %v", err)
-	}
-	if err := os.Remove(filepath.Join(dir, "change.md")); err != nil {
-		t.Fatalf("Remove change.md: %v", err)
-	}
-}
-
-func TestConversionPreCheckedBoxesAreCheckViolation(t *testing.T) {
-	repo := initCLIGitRepo(t)
-	folder := writeChangeFolder(t, repo, "20260727-prechecked", changeDoc(
-		"---\nchange: prechecked\ncreated: 2026-07-27\nbranch: prechecked\ntarget_release: 2.0.0\n---\n",
-		append(productSections(), executableSections()...)...,
-	))
-	commitAllChangeTest(t, repo, "docs: add legacy targeted change")
-	atomicConvertFolder(t, repo, "20260727-prechecked", "prechecked", true)
-	commitAllChangeTest(t, repo, "docs: convert with pre-checked box")
-
-	out, err := runChangeCheckJSON(t, repo, folder)
-	if err == nil || !findingsContain(out.Findings, "TASK-001-do-work.md") || !findingsContain(out.Findings, "checked task checkbox") {
-		t.Fatalf("err=%v findings=%v, want conversion violation naming TASK-001-do-work.md", err, out.Findings)
-	}
-}
-
-func TestConversionAllUncheckedPassesAndRealDogfoodIsCovered(t *testing.T) {
-	repo := initCLIGitRepo(t)
-	folder := writeChangeFolder(t, repo, "20260727-clean-convert", changeDoc(
-		"---\nchange: clean-convert\ncreated: 2026-07-27\nbranch: clean-convert\ntarget_release: 2.0.0\n---\n",
-		append(productSections(), executableSections()...)...,
-	))
-	commitAllChangeTest(t, repo, "docs: add legacy targeted change")
-	atomicConvertFolder(t, repo, "20260727-clean-convert", "clean-convert", false)
-	commitAllChangeTest(t, repo, "docs: atomic convert all unchecked")
-
-	out, err := runChangeCheckJSON(t, repo, folder)
-	if err != nil {
-		t.Fatalf("clean conversion check err=%v findings=%v", err, out.Findings)
-	}
-	if findingsContain(out.Findings, "checked task checkbox") {
-		t.Fatalf("findings=%v, want no conversion checkbox violation", out.Findings)
-	}
-
-	// This change's real dogfood conversion (acbea950) is the positive coverage
-	// target: grandfathered while INTENT-20260727-dogfood-conversion-manufactured-task-003-execution
-	// tracks remediation, check stays green and retention still treats it as replace.
-	repoRoot := filepath.Join("..", "..")
-	pilot := "docs/changes/20260726-change-work-model"
-	if _, err := os.Stat(filepath.Join(repoRoot, pilot, "change.json")); err != nil {
-		t.Skipf("change-work-model folder not present: %v", err)
-	}
-	findings, err := conversionPreCheckedFindings(repoRoot, pilot, commandOutput)
-	if err != nil {
-		t.Fatalf("real conversion findings err=%v", err)
-	}
-	if len(findings) != 0 {
-		t.Fatalf("real dogfood conversion findings=%v, want none (grandfathered acbea950)", findings)
-	}
-	// The dogfood commit lived on the change-work-model branch; the squash
-	// merge of PR #141 replaced it and branch deletion made it unreachable, so
-	// clones and CI checkouts legitimately do not have the object. The scanner
-	// leg below is history-dependent extra coverage, not the load-bearing
-	// assertion (the synthetic fixtures above are); skip it when the commit is
-	// absent rather than failing on a correct checkout.
-	const dogfoodConversion = "acbea95001f9187b154d095f4579225b7744fe1d"
-	if _, err := commandOutput(repoRoot, "git", "cat-file", "-e", dogfoodConversion+"^{commit}"); err != nil {
-		t.Skipf("dogfood conversion commit %s unreachable after squash merge: %v", dogfoodConversion[:8], err)
-	}
-	offending, err := conversionCommitCheckedTaskFiles(repoRoot, dogfoodConversion, pilot, commandOutput)
-	if err != nil {
-		t.Fatalf("inspect acbea950 tasks: %v", err)
-	}
-	if !findingsContain(offending, "TASK-003-check-and-projections.md") {
-		t.Fatalf("acbea950 offending=%v, want TASK-003 detected by the scanner (grandfather only suppresses the finding)", offending)
-	}
-}
-
-func TestTwoCommitConversionStillBlocksWithRetentionFinding(t *testing.T) {
-	repo := initCLIGitRepo(t)
-	folder := writeChangeFolder(t, repo, "20260727-two-step", changeDoc(
-		"---\nchange: two-step\ncreated: 2026-07-27\nbranch: two-step\ntarget_release: 2.0.0\n---\n",
-		append(productSections(), executableSections()...)...,
-	))
-	commitAllChangeTest(t, repo, "docs: add targeted legacy change")
-
-	if err := os.WriteFile(filepath.Join(folder, "change.json"), []byte(`{"change":"two-step","created":"2026-07-27","branch":"two-step","target_release":"2.0.0"}`), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	commitAllChangeTest(t, repo, "docs: add change.json keep-both")
-
-	if err := os.Remove(filepath.Join(folder, "change.md")); err != nil {
-		t.Fatal(err)
-	}
-	commitAllChangeTest(t, repo, "docs: retire change.md later")
-
-	deleted, err := deletedLineageChangesWithOutput(repo, commandOutput)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !findingsContain(deleted, "docs/changes/20260727-two-step/change.md") {
-		t.Fatalf("deleted=%v, want two-commit conversion retention finding for change.md", deleted)
-	}
-}
