+++ internal/cli/change_layout.go
@@ -20,17 +20,6 @@ const (
 	changeBriefFile         = "brief.md"
 )
 
-// changeRetargetEvent records a target_release mutation derived from unioned
-// history across change.json and change.md surfaces.
-type changeRetargetEvent struct {
-	Folder  string
-	Slug    string
-	From    string
-	To      string
-	Commit  string
-	Surface string
-}
-
 // assembleChangeNodeFromFolder builds a layout-agnostic node from a working-tree
 // Change folder. change.json presence selects the new layout; its absence falls
 // back to legacy change.md. A present but malformed change.json fails closed.
@@ -185,148 +174,4 @@ func changeFolderHasMachineSurface(folderAbs string) bool {
 		return true
 	}
 	return false
-}
-
-func deriveChangeCohorts(nodes []changeNode) map[string][]changeNode {
-	cohorts := map[string][]changeNode{}
-	for _, node := range nodes {
-		if node.TargetRelease == "" {
-			continue
-		}
-		cohorts[node.TargetRelease] = append(cohorts[node.TargetRelease], node)
-	}
-	for target, members := range cohorts {
-		sort.Slice(members, func(i, j int) bool { return members[i].Folder < members[j].Folder })
-		cohorts[target] = members
-	}
-	return cohorts
-}
-
-func changeFolderRelFromMachinePath(path string) string {
-	path = filepath.ToSlash(path)
-	if strings.HasSuffix(path, "/"+changeMachineFileJSON) || strings.HasSuffix(path, "/"+changeMachineFileLegacy) {
-		return filepath.ToSlash(filepath.Dir(path))
-	}
-	return path
-}
-
-// deriveChangeRetargetEvents unions change.json and change.md histories per
-// folder and returns target_release mutations (including removal-to-none).
-// Retargets are surfaced, never blocked.
-func deriveChangeRetargetEvents(rootPath string, outputCommand changeGitOutput) ([]changeRetargetEvent, error) {
-	output, err := outputCommand(rootPath, "git", "ls-tree", "-r", "--name-only", "HEAD", "--", "docs/changes")
-	if err != nil {
-		return nil, fmt.Errorf("inspect Change paths at HEAD: %w", err)
-	}
-	folders := map[string]bool{}
-	for _, path := range strings.Split(strings.TrimSpace(output), "\n") {
-		path = filepath.ToSlash(strings.TrimSpace(path))
-		base := filepath.Base(path)
-		if base == changeMachineFileJSON || base == changeMachineFileLegacy {
-			folders[filepath.ToSlash(filepath.Dir(path))] = true
-		}
-	}
-	// Also include folders that only exist in history (deleted) — retarget
-	// surfacing for current HEAD nodes is enough for TASK-001 consumers.
-	folderList := sortedKeys(folders)
-	var events []changeRetargetEvent
-	for _, folder := range folderList {
-		folderEvents, err := deriveFolderRetargetEvents(rootPath, folder, outputCommand)
-		if err != nil {
-			return nil, err
-		}
-		events = append(events, folderEvents...)
-	}
-	return events, nil
-}
-
-func deriveFolderRetargetEvents(rootPath, folder string, outputCommand changeGitOutput) ([]changeRetargetEvent, error) {
-	mdPath := filepath.ToSlash(filepath.Join(folder, changeMachineFileLegacy))
-	jsonPath := filepath.ToSlash(filepath.Join(folder, changeMachineFileJSON))
-	versions, err := loadUnionTargetHistory(rootPath, folder, mdPath, jsonPath, outputCommand)
-	if err != nil {
-		return nil, err
-	}
-	var events []changeRetargetEvent
-	prev := ""
-	havePrev := false
-	for _, version := range versions {
-		if !havePrev {
-			prev = version.Target
-			havePrev = true
-			continue
-		}
-		if version.Target == prev {
-			continue
-		}
-		events = append(events, changeRetargetEvent{
-			Folder:  folder,
-			Slug:    version.Slug,
-			From:    prev,
-			To:      version.Target,
-			Commit:  version.Commit,
-			Surface: version.Surface,
-		})
-		prev = version.Target
-	}
-	return events, nil
-}
-
-type changeTargetVersion struct {
-	Commit  string
-	Target  string
-	Slug    string
-	Surface string
-}
-
-func loadUnionTargetHistory(rootPath, folder, mdPath, jsonPath string, outputCommand changeGitOutput) ([]changeTargetVersion, error) {
-	commitsOutput, err := outputCommand(rootPath, "git", "rev-list", "--full-history", "--topo-order", "--reverse", "HEAD", "--", mdPath, jsonPath)
-	if err != nil {
-		return nil, fmt.Errorf("read %s target history: %w", folder, err)
-	}
-	commits := strings.Fields(commitsOutput)
-	var versions []changeTargetVersion
-	for _, commit := range commits {
-		jsonContent, jsonOK, err := readCommittedOptional(rootPath, commit, jsonPath, outputCommand)
-		if err != nil {
-			return nil, err
-		}
-		mdContent, mdOK, err := readCommittedOptional(rootPath, commit, mdPath, outputCommand)
-		if err != nil {
-			return nil, err
-		}
-		if !jsonOK && !mdOK {
-			continue
-		}
-		version := changeTargetVersion{Commit: commit}
-		if jsonOK {
-			meta := parseChangeJSON(jsonContent)
-			version.Target = meta.TargetRelease
-			version.Slug = meta.Change
-			version.Surface = changeMachineFileJSON
-		} else if mdOK {
-			fields, _ := changeFrontmatterFields(mdContent)
-			version.Target = changeFieldValue(fields, "target_release")
-			version.Slug = changeFieldValue(fields, "change")
-			version.Surface = changeMachineFileLegacy
-		}
-		versions = append(versions, version)
-	}
-	return versions, nil
-}
-
-func readCommittedOptional(rootPath, commit, path string, outputCommand changeGitOutput) (string, bool, error) {
-	treePath, err := outputCommand(rootPath, "git", "ls-tree", "--name-only", commit, "--", path)
-	if err != nil {
-		return "", false, fmt.Errorf("inspect %s at %s: %w", path, shortChangeCommit(commit), err)
-	}
-	treePath = filepath.ToSlash(strings.TrimSpace(treePath))
-	if treePath == "" {
-		return "", false, nil
-	}
-	content, err := outputCommand(rootPath, "git", "show", commit+":"+path)
-	if err != nil {
-		return "", false, fmt.Errorf("read %s at %s: %w", path, shortChangeCommit(commit), err)
-	}
-	return content, true, nil
 }
