--- internal/cli/change_plan_template.md
@@ -1,21 +1,0 @@
-<!-- plan.md is the optional technical route (the corpus's plan sense).
-     Accretive during shaping; approach churn here never expires a cohort receipt.
-     Criteria live in shape.md — do not relocate Verification Contract here. -->
-
-# Plan — [Change Title]
-
-## Approach
-
-[How the work lands — architecture, sequencing, compatibility.]
-
-## Placement
-
-[Where code and docs live; what stays out of scope for this route.]
-
-## Risks
-
-[Failure modes and the safe failure direction.]
-
-## Sequencing
-
-[Ordered slices that leave main coherent at each landing.]
