+++ internal/cli/cli_reference_test.go
@@ -34,7 +34,7 @@ func TestRunnerGenerateCLIReferenceWritesSkillNatively(t *testing.T) {
 		"## Journal Context (contract v2)",
 		"`loaf --help`",
 		"`loaf <command> --help`",
-		"`loaf config check --json`, `loaf state doctor --json`, `loaf change check --json`",
+		"`loaf config check --json`, `loaf state doctor --json`",
 		"active-truth read model",
 		"`project-synthesis`",
 		"`scoped-checkpoint`",
@@ -58,7 +58,6 @@ func TestRunnerGenerateCLIReferenceWritesSkillNatively(t *testing.T) {
 		"| Command | Purpose | Subcommands |",
 		"| `loaf build` | Build skill distributions for agent harnesses | — |",
 		"| `loaf config` | Validate and refresh project Loaf config | check |",
-		"| `loaf change` | Shape-first Change artifacts: git-canonical work context under docs/changes/ | init, check, list |",
 		"| `loaf journal` | Record and read the project-scoped journal (the durable record across all conversations) | log, recent, search, show, context, export, defer |",
 		"| `loaf state` | Manage native SQLite state | path, status, init, doctor, repair legacy-project-database,",
 		"repair journal-search",
