+++ internal/cli/delete_command_test.go
@@ -35,57 +35,6 @@ spec: SPEC-001
 	}
 }
 
-func TestRunnerSpecDeleteRemovesSpecAndDependents(t *testing.T) {
-	workingDir := realpath(t, t.TempDir())
-	stateHome := t.TempDir()
-	migrateSpecFixture(t, workingDir, stateHome)
-
-	var jsonOut bytes.Buffer
-	err := Runner{Stdout: &jsonOut, WorkingDir: workingDir, StateHome: stateHome}.Run([]string{"spec", "delete", "SPEC-001", "--yes", "--json"})
-	if err != nil {
-		t.Fatalf("spec delete --yes --json error = %v", err)
-	}
-	var result state.SpecDeleteResult
-	if err := json.Unmarshal(jsonOut.Bytes(), &result); err != nil {
-		t.Fatalf("json.Unmarshal(%q) error = %v", jsonOut.String(), err)
-	}
-	removed := map[string]int{}
-	for _, count := range result.Removed {
-		removed[count.Table] = count.Rows
-	}
-	if removed["specs"] != 1 || removed["aliases"] < 1 || removed["artifact_bodies"] < 1 {
-		t.Fatalf("removed = %#v, want specs=1 and dependent rows removed", removed)
-	}
-
-	// The spec is gone.
-	if err := (Runner{Stdout: &bytes.Buffer{}, WorkingDir: workingDir, StateHome: stateHome}).Run([]string{"spec", "show", "SPEC-001"}); err == nil {
-		t.Fatal("spec show after delete = nil error, want not-found failure")
-	}
-	// The linked task survives.
-	if err := (Runner{Stdout: &bytes.Buffer{}, WorkingDir: workingDir, StateHome: stateHome}).Run([]string{"task", "show", "TASK-001"}); err != nil {
-		t.Fatalf("task show after spec delete error = %v, want task preserved", err)
-	}
-}
-
-func TestRunnerSpecDeleteRequiresYes(t *testing.T) {
-	workingDir := realpath(t, t.TempDir())
-	stateHome := t.TempDir()
-	migrateSpecFixture(t, workingDir, stateHome)
-
-	var out bytes.Buffer
-	err := Runner{Stdout: &out, WorkingDir: workingDir, StateHome: stateHome}.Run([]string{"spec", "delete", "SPEC-001"})
-	if err == nil {
-		t.Fatal("spec delete without --yes = nil error, want refusal")
-	}
-	if !strings.Contains(err.Error(), "confirmation-required") {
-		t.Fatalf("error = %q, want confirmation-required message", err.Error())
-	}
-	// Nothing was deleted.
-	if err := (Runner{Stdout: &bytes.Buffer{}, WorkingDir: workingDir, StateHome: stateHome}).Run([]string{"spec", "show", "SPEC-001"}); err != nil {
-		t.Fatalf("spec show after refused delete error = %v, want spec intact", err)
-	}
-}
-
 func TestRunnerProjectDeleteCascades(t *testing.T) {
 	workingDir := realpath(t, t.TempDir())
 	stateHome := t.TempDir()
